Revert "Implemented sliding window attention to maintain KV cache only for the window size to enable infinite decoding. (#995)"
This reverts commit 67645d080314d7c374de50c336257e7b6d66fadc.

+++ axlearn/common/decoder.py
@@ -617,7 +617,7 @@ class Decoder(BaseLayer):
         cfg: Decoder.Config = self.config
         init_state, _ = self.transformer.init_states(
             time_step=None,
-            data=TensorSpec([batch_size, max_sequence_length, cfg.dim], dtype=jnp.float32),
+            data=TensorSpec([batch_size, max_sequence_length, cfg.dim]),
         )
         return dict(
             transformer_state=init_state,
